Extract SpotifySearcher class
