+++ src/main/java/chart/spotify/TrackSearchingAugmentor.java
@@ -7,7 +7,6 @@ import chart.postgres.raw.TrackRecord;
 import com.google.common.collect.ImmutableList;
 import com.wrapper.spotify.models.SimpleArtist;
 import com.wrapper.spotify.models.Track;
-import org.apache.commons.lang.StringUtils;
 
 import java.util.List;
 import java.util.Optional;
@@ -18,16 +17,17 @@ import java.util.stream.Collectors;
  *  In this case, we must search Postgres or Spotify for both the artist and the title.
  */
 public class TrackSearchingAugmentor implements SpotifyAugmentor {
-    private final SpotifyApi api;
     private final PostgresConnection connection;
+    private final SpotifySearcher spotifySearcher;
 
     public static TrackSearchingAugmentor create(SpotifyApi api, PostgresConnection connection) {
-        return new TrackSearchingAugmentor(api, connection);
+        SpotifySearcher spotifySearcher = new SpotifySearcher(api);
+        return new TrackSearchingAugmentor(connection, spotifySearcher);
     }
 
-    private TrackSearchingAugmentor(SpotifyApi api, PostgresConnection connection) {
-        this.api = api;
+    private TrackSearchingAugmentor(PostgresConnection connection, SpotifySearcher spotifySearcher) {
         this.connection = connection;
+        this.spotifySearcher = spotifySearcher;
     }
 
     @Override
@@ -59,6 +59,7 @@ public class TrackSearchingAugmentor implements SpotifyAugmentor {
     }
 
     private SpotifyChartEntry fetchFromSpotify(ChartEntry entry) {
+        // TODO handle cases where track was removed from Spotify
         String title = entry.title();
         String artist = entry.artist();
         Track track = getTrack(title, artist);
@@ -70,42 +71,9 @@ public class TrackSearchingAugmentor implements SpotifyAugmentor {
     }
 
     private Track getTrack(String title, String artist) {
-        List<Track> tracks = api.searchForTrack(title, artist);
+        Optional<Track> bestMatch = spotifySearcher.searchForTrack(title, artist);
 
-        Track closestMatch = null;
-        int closestDistance = Integer.MAX_VALUE;
-        for (Track item : tracks) {
-            boolean sameTitle = item.getName().equalsIgnoreCase(title);
-            boolean sameArtist = item.getArtists().stream().anyMatch(a -> a.getName().equalsIgnoreCase(artist));
-            if (sameTitle && sameArtist) {
-                return item;
-            }
-
-            if (sameArtist) {
-                int distance = StringUtils.getLevenshteinDistance(title, item.getName());
-                if (distance < closestDistance) {
-                    closestDistance = distance;
-                    closestMatch = item;
-                }
-            }
-        }
-
-        if (closestMatch != null && closestDistance < 5) {
-            return closestMatch;
-        }
-
-        throw new IllegalStateException(String.format(
-                "Couldn't find exact matches for track %s by %s - potential matches were:\n%s", title, artist,
-                printSearchResults(tracks)));
+        return bestMatch.get();
     }
 
-    private String printSearchResults(List<Track> tracks) {
-        return tracks.stream()
-                .map(this::printTrack)
-                .collect(Collectors.joining("\n"));
-    }
-
-    private String printTrack(Track track) {
-        return track.getArtists().get(0).getName() + " - " + track.getName();
-    }
 }
